Make pid_t print and scanning more portable
We can't assume pid_t is 32 or 64 bit, so make all print and scanning
for pid_t a intmax_t and cast appropriately. Use PRIdMAX and SCNdMAX
for intmax_t values.

+++ stress-mq.c
@@ -98,8 +98,8 @@ static int stress_mq(const stress_args_t *args)
 	}
 	sz = mq_size;
 
-	(void)snprintf(mq_name, sizeof(mq_name), "/%s-%i-%" PRIu32,
-		args->name, args->pid, args->instance);
+	(void)snprintf(mq_name, sizeof(mq_name), "/%s-%" PRIdMAX "-%" PRIu32,
+		args->name, (intmax_t)args->pid, args->instance);
 	fp = fopen("/proc/sys/fs/mqueue/msg_default", "r");
 	if (fp) {
 		if (fscanf(fp, "%20d", &max_sz) != 1)
