+++ stress-seal.c
@@ -74,8 +74,9 @@ static int stress_seal(const stress_args_t *args)
 		uint8_t *ptr;
 		char buf[page_size];
 
-		(void)snprintf(filename, sizeof(filename), "%s-%d-%" PRIu32 "-%" PRIu32,
-			args->name, args->pid, args->instance, stress_mwc32());
+		(void)snprintf(filename, sizeof(filename), "%s-%" PRIdMAX "-%" PRIu32 "-%" PRIu32,
+			args->name, (intmax_t)args->pid,
+			args->instance, stress_mwc32());
 
 		fd = shim_memfd_create(filename, MFD_ALLOW_SEALING);
 		if (fd < 0) {
